ShareJS with JSON type.
This is our first cut at this ... it is persisting the json
object to the database seemingly fine, and if you reload the
page in your browser or another it repopulates the editor
contents from what is in the database.

The issue right now is that the onInsert and onRemove functions
don't seem to be getting called, and there is some weird loop
that happens between check() and the delete functionality.