CodeMirror content is synching!
Great success!  The trick was that a lot of the shareJS documentation
is actually out of date ... looking at the API in the code ended up
being the most helpful thing.

The events that we want to listen for are 'insert' and 'delete', and adding an event listener also actually requires supplying the json path that you wish to listen on as well.  For the two listeners we have right now, these are just solely listening on the document's text attribute.