Created Chat model using thinky.
The chat logs are now stored in rethinkDB, thanks to thinky!  The route to collect the last 50 chat logs is also now using thinky and a rethinkDB query.

Updated a lot of the terminology to remove references to pad, now calling that roomId and calling 'client' 'authorId'.

Once we have the author model being handled by thinky, we should remove the color attribute from the messages and do a join with the authors before we return the result.

Removed Elasticsearch from package.json and the README, as it is no longer being used.