Adding tests for chat model.  #6
Added tests for validation and creation of new Chat records.  This led me to realize that the chat.color field was not required, so updated that in /models/Chat.js.

Also, rather than having the chat socket handler be responsible for generating Chat ids, thinky provides a handle to rethinkdashdb and its .uuid() function.  So, now in the model we just default the id field to use the result of that function if one is not provided.

Tests can be run with 'npm test'.  RethinkDB configuration is available in config/env/test.json.  You must have a database named 'test' in your RethinkDB instance or it won't work.