JSON Sync working with paste and multi-line delete.
The code is not the prettiest thing in the world, but it is working!  Tested with multiple users deleting multiple lines at a time, pasting in multiple lines, and selecting and then pasting multiple lines over the multi-line selection.

Also, added license to the top of the file, and gave credit to the share-codemirror repository.

Still todo: see if we can abstract this away from our implementation and make it usable for other situations.  If we can, it would probably make sense to have this as its own node package.