Created Author model, now saving authors in DB.
Used thinky to create an Author model which just contains the author ID and the color that their messages / modified editor lines should have.

Created a thinky helper to simplify creating the thinky instance.  The thinky documentation also indicates that models are not shared between instances, so it would be pointless to create multiple instances anyway.

Updated the chat socket connection handler to use the Author models.