Implemented basic replay functionality.  #7
Added a button (which we may want to move) that, when clicked, fetches all operations for the current document, and applies each operation one-by-one, drawing and then waiting with setTimeout after each application.

Playback is done within a separate CodeMirror instance ... I don't see a way around this, unless there is a way to detach the share-codemirror-json addon from the real CodeMirror instance, or at least disable it until the user is done using the replay.

Currently just drawing timestamps and text, not currently including chat history.  Timestamps break unless the changes are played through up to the current point.  We definitely need to fix that.

There is no way to go back and forth or choose a spot in the history right now.  We would definitely want to include that, and could also include some mechanism for controlling the playback speed pretty easily.