Updating the route to append text to a document.
Currently, this is just adding it to the head of the document, so it does not need to bother with updating the lines metadata, but eventually we want to have it appending new lines to the end of the document, in which case it will need to update that.

Additionally, invoking this route currently breaks any open connections to the affected Room, because their shareJS information is now out of date.

To fix that, we need to either be submitting this into the shareJS browserchannel so that all clients are notified of the update, or we need to handle in on the client side that after the initial discrepancy, it grabs the latest to rectify itself and get it working again.