Fixing broken timestamp hover.
Showed invalid date when the user hovered over a message that they had just typed.  We were specifying the date as ISO8601, which was correct for the dates coming from the server, but not for the date that had just been created on the client.

Moment does not appear to need us to specify that, though, so just removing it fixes the issue.